further requirements for the k6 performance tests
